Fix a spinlock in invokeOnCompletion with onCancelling = true
Before this change, when children were prohibited from adding
themselves to the list, cancellation handlers were also prohibited
from doing so. This was plain incorrect, because a list that's
closed for new children could still get cancelled later, and also
because being closed for new children happened before advancing the
state to the final one.